Add fort3 replacement script
Since the input to this function will have to be quite a lot more complicated
than the input to the corresponding fort1 replacement script, a new class is
made to handle it. Some logic is attached to the class to make sure it is semi
self consistent (in that e.g. the same property cannot both be added as a new
property *and* replace an existing one at the same time).